[refactoring] will move sortItems to sort-menu
