+++ src/main.js
@@ -12,27 +12,10 @@ import {createNewComment} from './view/new-comment.js';
 import {generateFilm} from './mock/mock-film.js';
 import {generateComment} from './mock/mock-comment.js';
 import {getRandomInteger, sortFilmByComments, sortFilmByRating} from './utils';
-import {NavigationTypes, SortTypes, FilmListTypes} from './utils/const.js';
+import {NavigationTypes, FilmListTypes} from './utils/const.js';
 
 const CARDS_LOAD_STEP = 5;
 const CARDS_EXTRA_LOAD_STEP = 2;
-const sortItems = [
-  {
-    type: SortTypes.DEFAULT,
-    text: 'Sort by default',
-    isActive: true,
-  },
-  {
-    type: SortTypes.DATE,
-    text: 'Sort by date',
-    isActive: false,
-  },
-  {
-    type: SortTypes.RATING,
-    text: 'Sort by rating',
-    isActive: false,
-  },
-];
 
 const headerElement = document.querySelector('.header');
 const mainElement = document.querySelector('.main');
@@ -72,7 +55,7 @@ const renderFilmCards = (filmList, loadStep, target, startPosition=0) => {
 
 render(headerElement, createUserProfile(watchedFilmsCount));
 render(mainElement, createMainNavigation(navigationStatistics));
-render(mainElement, createSortMenu(sortItems));
+render(mainElement, createSortMenu());
 render(mainElement, createFilmsBoard(filmsData));
 
 const filmBoard = document.querySelector('.films');
